+++ index.js
@@ -23,3 +23,4 @@ exports.new = function(Constructor) {
 exports.noop = function() {}
 exports.identity = function(value) { return value }
 exports.constant = function(value) { return exports.identity.bind(null, value) }
+exports.clone = function(obj) { return obj ? exports.assign({}, obj)  : obj }