+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Adds `Overstrike.noop`.
 - Adds `Overstrike.identity`.
 - Adds `Overstrike.constant`.
+- Adds `Overstrike.clone`.
 
 ## 0.1.337 (Aug 4, 2014)
 - Adds `Overstrike.assign`.